Move crypto to its own place
